requirements.txt
selenium
pyotp
Pillow
python-dotenv